+++ ICD.Connect.Routing/RoutingGraphs/AbstractRoutingGraph.cs
@@ -126,6 +126,31 @@ namespace ICD.Connect.Routing.RoutingGraphs
 		/// <summary>
 		/// Finds the destinations that the source is actively routed to.
 		/// </summary>
+		/// <param name="source"></param>
+		/// <param name="type"></param>
+		/// <param name="signalDetected">When true skips inputs where no video is detected.</param>
+		/// <param name="inputActive"></param>
+		/// <exception cref="ArgumentNullException"></exception>
+		/// <returns>The sources</returns>
+		public abstract IEnumerable<EndpointInfo> GetActiveDestinationEndpoints(ISource source, eConnectionType type,
+		                                                                        bool signalDetected, bool inputActive);
+
+		/// <summary>
+		/// Finds the destinations that the source is actively routed to.
+		/// </summary>
+		/// <param name="sourceOutput"></param>
+		/// <param name="type"></param>
+		/// <param name="signalDetected">When true skips inputs where no video is detected.</param>
+		/// <param name="inputActive"></param>
+		/// <exception cref="ArgumentNullException"></exception>
+		/// <returns>The sources</returns>
+		public abstract IEnumerable<EndpointInfo> GetActiveDestinationEndpoints(EndpointInfo sourceOutput,
+		                                                                        eConnectionType type, bool signalDetected,
+		                                                                        bool inputActive);
+
+		/// <summary>
+		/// Finds the destinations that the source is actively routed to.
+		/// </summary>
 		/// <param name="sourceControl"></param>
 		/// <param name="sourceOutput"></param>
 		/// <param name="type"></param>